chore: move search domain to search package
