+++ internal/api/api.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/newrelic/go-agent/v3/newrelic"
+	"github.com/odpf/stencil/core/search"
 	"github.com/odpf/stencil/domain"
 	stencilv1beta1 "github.com/odpf/stencil/proto/odpf/stencil/v1beta1"
 	"google.golang.org/grpc/health/grpc_health_v1"
@@ -21,10 +22,10 @@ type API struct {
 	grpc_health_v1.UnimplementedHealthServer
 	namespace domain.NamespaceService
 	schema    domain.SchemaService
-	search    domain.SearchService
+	search    search.SearchService
 }
 
-func NewAPI(namespace domain.NamespaceService, schema domain.SchemaService, search domain.SearchService) *API {
+func NewAPI(namespace domain.NamespaceService, schema domain.SchemaService, search search.SearchService) *API {
 	return &API{
 		namespace: namespace,
 		schema:    schema,
