+++ internal/store/postgres/repository.go
@@ -8,6 +8,7 @@ import (
 	"github.com/georgysavva/scany/pgxscan"
 	"github.com/jackc/pgconn"
 	"github.com/jackc/pgx/v4"
+	"github.com/odpf/stencil/core/search"
 	"github.com/odpf/stencil/domain"
 	"github.com/odpf/stencil/internal/store"
 )
@@ -151,14 +152,14 @@ func (r *Store) DeleteVersion(ctx context.Context, ns string, sc string, version
 	return wrapError(err, "delete version")
 }
 
-func (r *Store) Search(ctx context.Context, req *domain.SearchRequest) ([]*domain.SearchHits, error) {
-	var searchHits []*domain.SearchHits
+func (r *Store) Search(ctx context.Context, req *search.SearchRequest) ([]*search.SearchHits, error) {
+	var searchHits []*search.SearchHits
 	err := pgxscan.Select(ctx, r.db, &searchHits, searchAllQuery, req.NamespaceID, req.SchemaID, req.VersionID, req.Query)
 	return searchHits, err
 }
 
-func (r *Store) SearchLatest(ctx context.Context, req *domain.SearchRequest) ([]*domain.SearchHits, error) {
-	var searchHits []*domain.SearchHits
+func (r *Store) SearchLatest(ctx context.Context, req *search.SearchRequest) ([]*search.SearchHits, error) {
+	var searchHits []*search.SearchHits
 	err := pgxscan.Select(ctx, r.db, &searchHits, searchLatestQuery, req.NamespaceID, req.SchemaID, req.Query)
 	return searchHits, err
 }
